+++ graph_cast/onto.py
@@ -1,0 +1,12 @@
+from enum import Enum
+from types import MappingProxyType
+
+
+class InputType(str, Enum):
+    JSON = "json"
+    TABLE = "table"
+
+
+InputTypeFileExtensions = MappingProxyType(
+    {InputType.JSON: (InputType.JSON,), InputType.TABLE: ("csv",)}
+)