+++ run/arango/ingest_table.py
@@ -10,7 +10,7 @@ logger = logging.getLogger(__name__)
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
 
-    parser.add_argument("--path", type=str, help="path to csv datafiles")
+    parser.add_argument("--path", type=str, help="path to table datafiles")
 
     parser.add_argument(
         "-f",
@@ -91,5 +91,5 @@ if __name__ == "__main__":
         limit_files=args.limit_files,
         max_lines=args.max_lines,
         clean_start=args.clean_start,
-        n_thread=args.n_thread,
+        n_threads=args.n_thread,
     )